move chain selector
