+++ src/components/WalletSuite/ConnectedWallet/Details/index.tsx
@@ -7,9 +7,9 @@ import { WalletState, useSetWallet } from "contexts/WalletContext";
 import { logger } from "helpers";
 import { GENERIC_ERROR_MESSAGE } from "constants/common";
 import { appRoutes } from "constants/routes";
+import ChainSelector from "../../../donation/Steps/Donater/Form/ChainSelector";
 import Address from "./Address";
 import Balances from "./Balances";
-import ChainSelector from "./ChainSelector";
 import Favourites from "./Favourites";
 import MobileTitle from "./MobileTitle";
 import MyEndowments from "./MyEndowments";
